refactor: adhered to correct directory structure.
